COMMERCE-4667 commerce-product-api commerce-product-service AactionHelper splitted into interface and impl
